Major changes
Several changes to the sound replacement script: overhauled tempo, automation, and subtrack editing, hopefully for more accurate results.

Ended up switching to a custom build of wwiser as well that makes use of mentioning additional parameters in the Wwise banks, which can be used to further simplify sounbank editing. If there is enough interest, I can try to properly fork wwiser for showing the uncompiled code.